creates components for client and styles
